+++ lib/src/playx_core.dart
@@ -6,9 +6,11 @@ import '../playx_core.dart';
 /// used setup and disposing required dependencies needed for playx_core.
 abstract class PlayXCore {
   /// used initialize and install dependencies needed for playx_core.
-  static Future<void> bootCore(
-      {SecurePrefsSettings securePrefsSettings =
-          const SecurePrefsSettings()}) async {
+  static Future<void> bootCore({
+    PlayxSecurePrefsSettings securePrefsSettings =
+        const PlayxSecurePrefsSettings(),
+    PlayxEnvSettings? envSettings,
+  }) async {
     /// inject SharedPreferences
     final prefs = await SharedPreferences.getInstance();
     Get.put<SharedPreferences>(prefs, permanent: true);
@@ -21,6 +23,14 @@ abstract class PlayXCore {
       mOptions: securePrefsSettings.macOsOptions,
     );
     Get.put<FlutterSecureStorage>(securePrefs, permanent: true);
+    if (envSettings != null) {
+      await PlayxEnv.load(
+        fileName: envSettings.fileName,
+        parser: envSettings.parser,
+        mergeWith: envSettings.mergeWith,
+        isOptional: envSettings.isOptional,
+      );
+    }
   }
 
   ///Dispose PlayxCore